requirements.txt
sentence-transformers
pandas
scikit-learn
torch
transformers
tqdm
groq
python-dotenv
nltk
datasets
